Fix frontend routes and rename MojiTiketi to MyTickets
