+++ Project/frontend/src/App.jsx
@@ -5,7 +5,7 @@ import Home from "./pages/Home"
 import Login from "./pages/Login"
 import Dashboard from "./pages/Dashboard"
 import Tickets from "./pages/Tickets"
-import MojiTiketi from "./pages/MojiTiketi"
+import MyTickets from "./pages/MyTickets"
 
 function App() {
   return (
@@ -16,14 +16,14 @@ function App() {
           <Route path="/login" element={<Login />} />
 
           {/* US-11: Zaštićena ruta — redirect na /login ako korisnik nije prijavljen */}
-          <Route path="/mojitiketi" element={
-            <ProtectedRoute><MojiTiketi /></ProtectedRoute>
+          <Route path="/mytickets" element={
+            <ProtectedRoute><MyTickets /></ProtectedRoute>
           } />
 
           <Route path="/dashboard" element={
             <ProtectedRoute><Dashboard /></ProtectedRoute>
           } />
-          <Route path="/Tickets" element={
+          <Route path="/tickets" element={
             <ProtectedRoute><Tickets /></ProtectedRoute>
           } />
           <Route path="*" element={<Navigate to="/" replace />} />
